+++ web/src/MissionsRoute.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MissionsRoute: React.FC = () => {
+  return <div>Missions page coming soon!</div>;
+};
+ 
+export default MissionsRoute;